Split image generation executor helpers
