+++ frontend/src/server/images/image-generation-receipts.ts
@@ -1,0 +1,95 @@
+import { query } from '@/lib/db';
+import type { BillingProductKey, JobSurface } from '@/types/billing';
+import type { PricingSnapshot } from '@/types/engines';
+
+export type PendingReceipt = {
+  userId: string;
+  amountCents: number;
+  currency: string;
+  description: string;
+  jobId: string;
+  surface: JobSurface;
+  billingProductKey: BillingProductKey | null;
+  snapshot: unknown;
+  applicationFeeCents: number | null;
+  vendorAccountId: string | null;
+  stripePaymentIntentId?: string | null;
+  stripeChargeId?: string | null;
+};
+
+export function buildReceiptSnapshot(pricing: PricingSnapshot): Record<string, unknown> {
+  const snapshot: Record<string, unknown> = {
+    totalCents: pricing.totalCents,
+    currency: pricing.currency,
+  };
+
+  const discountCandidate = (pricing as {
+    discount?: { amountCents?: number; percentApplied?: number; label?: string };
+  }).discount;
+  if (discountCandidate && typeof discountCandidate.amountCents === 'number' && discountCandidate.amountCents > 0) {
+    snapshot.discount = {
+      amountCents: discountCandidate.amountCents,
+      percentApplied: discountCandidate.percentApplied ?? null,
+      label: discountCandidate.label ?? null,
+    };
+  }
+
+  const taxesCandidate = (pricing as { taxes?: Array<{ amountCents?: number; label?: string }> }).taxes;
+  if (Array.isArray(taxesCandidate)) {
+    const taxes = taxesCandidate
+      .filter((tax) => tax && typeof tax.amountCents === 'number' && tax.amountCents > 0)
+      .map((tax) => ({
+        amountCents: tax.amountCents!,
+        label: tax.label ?? null,
+      }));
+    if (taxes.length) {
+      snapshot.taxes = taxes;
+    }
+  }
+
+  return snapshot;
+}
+
+export async function recordRefundReceipt(receipt: PendingReceipt, label: string, priceOnly: boolean) {
+  try {
+    await query(
+      `INSERT INTO app_receipts (
+         user_id,
+         type,
+         amount_cents,
+         currency,
+         description,
+         job_id,
+         surface,
+         billing_product_key,
+         pricing_snapshot,
+         application_fee_cents,
+         vendor_account_id,
+         stripe_payment_intent_id,
+         stripe_charge_id,
+         platform_revenue_cents,
+         destination_acct
+       )
+       VALUES ($1,'refund',$2,$3,$4,$5,$6,$7,$8::jsonb,$9,$10,$11,$12,$13,$14)
+       ON CONFLICT DO NOTHING`,
+      [
+        receipt.userId,
+        receipt.amountCents,
+        receipt.currency,
+        label,
+        receipt.jobId,
+        receipt.surface,
+        receipt.billingProductKey,
+        JSON.stringify(receipt.snapshot),
+        priceOnly ? null : 0,
+        priceOnly ? null : receipt.vendorAccountId,
+        receipt.stripePaymentIntentId ?? null,
+        receipt.stripeChargeId ?? null,
+        priceOnly ? null : 0,
+        priceOnly ? null : receipt.vendorAccountId,
+      ]
+    );
+  } catch (error) {
+    console.warn('[images] failed to record refund receipt', error);
+  }
+}
