+++ tests/image-generation-server-architecture.test.ts
@@ -10,6 +10,8 @@ const referenceNormalizationPath = join(root, 'frontend/src/server/images/image-
 const initialJobPath = join(root, 'frontend/src/server/images/image-initial-job.ts');
 const errorPath = join(root, 'frontend/src/server/images/image-generation-error.ts');
 const providerPayloadPath = join(root, 'frontend/src/server/images/image-provider-payload.ts');
+const receiptsPath = join(root, 'frontend/src/server/images/image-generation-receipts.ts');
+const settingsSnapshotPath = join(root, 'frontend/src/server/images/image-generation-settings-snapshot.ts');
 
 const executorSource = readFileSync(executorPath, 'utf8');
 const existingJobResponseSource = readFileSync(existingJobResponsePath, 'utf8');
@@ -17,6 +19,8 @@ const referenceNormalizationSource = readFileSync(referenceNormalizationPath, 'u
 const initialJobSource = readFileSync(initialJobPath, 'utf8');
 const errorSource = readFileSync(errorPath, 'utf8');
 const providerPayloadSource = readFileSync(providerPayloadPath, 'utf8');
+const receiptsSource = readFileSync(receiptsPath, 'utf8');
+const settingsSnapshotSource = readFileSync(settingsSnapshotPath, 'utf8');
 
 test('image generation executor delegates focused server helpers', () => {
   assert.ok(existsSync(existingJobResponsePath), 'existing image job response helpers should live in a focused module');
@@ -24,11 +28,15 @@ test('image generation executor delegates focused server helpers', () => {
   assert.ok(existsSync(initialJobPath), 'atomic initial image job creation should live in a focused module');
   assert.ok(existsSync(errorPath), 'image generation execution error should live in a focused module');
   assert.ok(existsSync(providerPayloadPath), 'provider payload parsing should live in a focused module');
+  assert.ok(existsSync(receiptsPath), 'image generation receipt helpers should live in a focused module');
+  assert.ok(existsSync(settingsSnapshotPath), 'image settings snapshot helpers should live in a focused module');
   assert.match(executorSource, /from '\.\/existing-image-job-response'/);
   assert.match(executorSource, /from '\.\/image-reference-normalization'/);
   assert.match(executorSource, /from '\.\/image-initial-job'/);
   assert.match(executorSource, /from '\.\/image-generation-error'/);
   assert.match(executorSource, /from '\.\/image-provider-payload'/);
+  assert.match(executorSource, /from '\.\/image-generation-receipts'/);
+  assert.match(executorSource, /from '\.\/image-generation-settings-snapshot'/);
   assert.match(executorSource, /export \{ buildResponseFromExistingJob \} from '\.\/existing-image-job-response'/);
   assert.match(executorSource, /export \{ ImageGenerationExecutionError \} from '\.\/image-generation-error'/);
 });
@@ -49,9 +57,12 @@ test('image generation executor does not regain extracted server ownership', () 
   assert.doesNotMatch(executorSource, /function buildCharacterReferencePrompt\(/, 'character reference prompt building belongs in image-provider-payload.ts');
   assert.doesNotMatch(executorSource, /function extractImages\(/, 'provider image extraction belongs in image-provider-payload.ts');
   assert.doesNotMatch(executorSource, /function parseRequestId\(/, 'provider request id parsing belongs in image-provider-payload.ts');
+  assert.doesNotMatch(executorSource, /function buildReceiptSnapshot\(/, 'receipt snapshots belong in image-generation-receipts.ts');
+  assert.doesNotMatch(executorSource, /async function recordRefundReceipt\(/, 'refund receipt writes belong in image-generation-receipts.ts');
+  assert.doesNotMatch(executorSource, /function buildDefaultSettingsSnapshot\(/, 'default settings snapshots belong in image-generation-settings-snapshot.ts');
 
   const lineCount = executorSource.split('\n').length;
-  assert.ok(lineCount <= 1110, `image generation executor should stay below 1110 lines after provider payload extraction, got ${lineCount}`);
+  assert.ok(lineCount <= 970, `image generation executor should stay below 970 lines after receipt and settings extraction, got ${lineCount}`);
 });
 
 test('existing image job response module exposes the expected contract', () => {
@@ -75,4 +86,8 @@ test('existing image job response module exposes the expected contract', () => {
   assert.match(providerPayloadSource, /export function buildCharacterReferencePrompt/);
   assert.match(providerPayloadSource, /export function extractImages/);
   assert.match(providerPayloadSource, /export function parseRequestId/);
+  assert.match(receiptsSource, /export type PendingReceipt/);
+  assert.match(receiptsSource, /export function buildReceiptSnapshot/);
+  assert.match(receiptsSource, /export async function recordRefundReceipt/);
+  assert.match(settingsSnapshotSource, /export function buildDefaultSettingsSnapshot/);
 });
